Generate random vector properly
Previously, the procedure generate_random_vector_in_unit_sphere()
generated a single random number and assigned this random number to each
component of the vector. This led to issues with diffusion. Now, the
procedure generates a random number for each component of the vector
individually with separate procedure generate_random_vector().